address model added

+++ Angular_Template_Forms/src/app/Component/User/user-list/user-list.component.ts
@@ -6,6 +6,7 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { SearchUsersPipe } from '../../../Pipes/search-users.pipe';
 import { HomeComponent } from '../../home/home.component';
+import { Task, TaskService } from '../../../Service/task.service';
 
 @Component({
   selector: 'app-user-list',
@@ -17,8 +18,13 @@ import { HomeComponent } from '../../home/home.component';
 export class UserListComponent {
   Users:User[] = [];
   SearchText:string = "";
+  Tasks:Task[] = [];
 
-  constructor(private userservice:UserService , private router:Router ,private toastr: ToastrService){}
+  constructor(private userservice:UserService , private router:Router ,private toastr: ToastrService , private taskservice:TaskService){
+    this.taskservice.getTask().subscribe((data) => {
+      this.Tasks = data;
+    })
+  }
 
   ngOnInit(): void {
     this.listUsers();
@@ -33,16 +39,26 @@ export class UserListComponent {
   }
 
   DeleteUser(id:number){
-    if(confirm("Do you want to delete?")){
-      this.userservice.deleteUser(id).subscribe((data) => {
-        this.toastr.success("User Deleted Successfully.." , "" , {
-          positionClass:"toast-top-right",
-          progressBar:true,
-          timeOut:2000
-        })
-        this.listUsers();
+    const task = this.Tasks.find(t => t.userId == id);
+    if(task){
+      this.toastr.warning("you can't delete this User." , "" , {
+        positionClass:"toast-top-right",
+        progressBar:true,
+        timeOut:2000
       })
+    }else{
+      if(confirm("Do you want to delete?")){
+        this.userservice.deleteUser(id).subscribe((data) => {
+          this.toastr.success("User Deleted Successfully.." , "" , {
+            positionClass:"toast-top-right",
+            progressBar:true,
+            timeOut:2000
+          })
+          this.listUsers();
+        })
+      }
     }
+    
   }
 
   listUsers(){
